% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaborativedatascience-package.R
\docType{package}
\name{collaborativedatasciencepackage-package}
\alias{collaborativedatasciencepackage}
\alias{collaborativedatasciencepackage-package}
\title{collaborativedatascience-package}
\description{
A package for analyzing benthic cover data from coral reef ecosystems,
including data wrangling, visualization, ordination, and modeling tools
developed for a collaborative data science project.
}
\details{
The collaborativedatasciencepackage provides functions and datasets
to support ecological analysis of coral reef benthic communities.
Key features include:

\itemize{
\item Data preparation and wrangling of quadrat-level observations
\item Functional group aggregation of benthic categories
\item Multivariate ordination (PCoA)
\item Visualization utilities for benthic community data
\item Modeling tools for coral–macroalgae relationships
}
}
\section{Package options}{

This package does not currently define any global options.
}

\author{
Carly R. Dempsey
}
\keyword{internal}
