% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Metadata for benthic quadrat images}
\format{
A data frame with the following columns:
\describe{
\item{Image.name}{Image filename}
\item{Site}{Sampling site code}
\item{Date}{Date of image collection}
\item{Transect}{Transect identifier}
\item{Quadrat}{Quadrat number within transect}
}
}
\source{
Coral reef field surveys
}
\usage{
data(metadata)
}
\description{
This dataset contains metadata for quadrat photographs collected along
transects across multiple coral reef habitats.
}
\keyword{datasets}
